refactor : ConsoleUILottoApplication 코드를 controller로 옮김
