+++ src/main/java/lotto/ConsoleUILottoApplication.java
@@ -1,46 +1,10 @@
 package lotto;
 
-import static lotto.view.ConsoleInputView.*;
-import static lotto.view.ConsoleOutputView.*;
-
 import lotto.controller.LottoController;
-import lotto.domain.lotto.Lotto;
-import lotto.domain.lotto.LottoParser;
-import lotto.domain.lotto.Lottos;
-import lotto.domain.lottomoney.InvalidLottoMoneyException;
-import lotto.domain.lottomoney.LottoMoney;
-import lotto.domain.lottonumber.LottoNumber;
-import lotto.domain.result.LottoRankRecord;
-import lotto.domain.result.WinningLotto;
 
 public class ConsoleUILottoApplication {
 	public static void main(String[] args) {
 		LottoController lottoController = new LottoController();
-		LottoMoney inputLottoMoney = receiveInputMoney();
-		int numberOfLotto = inputLottoMoney.calculateNumberOfLotto();
-		printPurchaseCompleteMessage(numberOfLotto);
-
-		Lottos lottos = lottoController.purchaseLotto(numberOfLotto);
-		printPurchasedLotto(lottos);
-
-		Lotto inputWinningLotto = new Lotto(LottoParser.parser(inputWinningLottoNumber()));
-		LottoNumber inputBonusNumber = LottoNumber.valueOf(inputBonusLottoNumber());
-		WinningLotto winningLotto = new WinningLotto(inputWinningLotto, inputBonusNumber);
-
-		LottoRankRecord lottoRankCount = new LottoRankRecord(lottos, winningLotto);
-		printStatisticsMessage();
-		printWinningResult(lottoRankCount.getLottoRankCount());
-
-		int winningRatio = lottoController.calculateWinningRatio(lottoRankCount.getLottoRankCount(), inputLottoMoney);
-		printWinningRatio(winningRatio);
-	}
-
-	private static LottoMoney receiveInputMoney() {
-		try {
-			return new LottoMoney(inputMoney());
-		} catch (InvalidLottoMoneyException ime) {
-			printExceptionMessage(ime.getMessage());
-			return receiveInputMoney();
-		}
+		lottoController.run();
 	}
 }
